requirements.txt
# --- Core App & UI ---
customtkinter==5.2.2
Pillow==11.2.1
darkdetect==0.8.0
pyperclip==1.9.0

# --- Web Scraping & Data Handling ---
playwright==1.52.0
beautifulsoup4==4.13.4
requests==2.32.4
lxml==5.4.0

# --- RAG Engine: LlamaIndex & Integrations ---
llama-index==0.10.64
llama-index-llms-ollama==0.1.3
llama-index-embeddings-huggingface==0.2.0
llama-index-vector-stores-faiss==0.1.2

# --- Vector Store (stable, CPU-based) ---
faiss-cpu==1.8.0

# --- Machine Learning Backbone (PyTorch for CUDA 12.1) ---
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121

# --- Core Transformers & Embedding Models ---
transformers==4.43.3
sentence-transformers==2.7.0

# --- Utilities ---
python-dotenv==1.0.1
tqdm==4.67.1